#35 Rename circuit package to circuits
This fixes an issue with the API reference not properly covering all aspects of the circuit package, due to it getting confused with the circuit module within the circuit package.
