+++ src/manim_eng/__init__.py
@@ -1,7 +1,7 @@
 """Plugin for Manim Community Edition to help with drawing engineering diagrams."""
 
 from ._config import *
-from .circuit import *
+from .circuits import *
 from .components import *
 from .logo import *
 from .units import *
