+++ src/manim_eng/circuits/base/wire.py
